Move verification of binary digest to verifier
